neuen Server Ordner angelegt um Server aus App.js zu bringen, Unnötige imports gelöscht, unter functions eine Connection.js angelegt um onConnectivityChange aus App.js herauszunehmen und entsprechend alle imports angepasst.
