+++ AVARE-App/src/server/server.js
@@ -1,0 +1,2 @@
+//Constant Adress of the Server
+export const SERVER = 'http://193.196.36.83:8443' // IP + Port of the host, or  http://localhost:8443 for testing on pc
